+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ CHANGELOG
 - Removed facts plugin. Now at https://github.com/shaunduncan/helga-facts (pypi helga-facts)
 - Removed giphy plugin. Now at https://github.com/shaunduncan/helga-giphy (pypi helga-giphy)
 - Removed icanhazascii plugin. Now at https://github.com/shaunduncan/helga-icanhazascii (pypi helga-icanhazascii)
+- Removed jira plugin. Now at https://github.com/shaunduncan/helga-jira (pypi helga-jira)
 - Removed loljava plugin. Now at https://github.com/shaunduncan/helga-loljava (pypi helga-loljava)
 - Removed no_more_olga plugin. Now at https://github.com/shaunduncan/helga-no-more-olga (pypi helga-no-more-olga)
 - Removed oneliner plugin. Now at https://github.com/shaunduncan/helga-oneliner (pypi helga-onliner)